requirements.txt
setuptools==58.0.4
wheel==0.37.0
numpy==1.20.3
pandas==1.3.4
pyspark==3.2.1
scikit-learn==0.24.2
pyyaml==6.0
databricks-cli==0.16.6
dbx==0.6.10
GitPython==3.1.27
delta-spark==1.2.1
python-dotenv==0.20.0
